// src/app/models/addInfo.ts
export class AddInfo {
    id: string;
    name:string;
    companyName:string;
    photo: string;
    category:string;
    description:string;
    teleLink:string;
    facebookLink: string;
    createdAt:string;
    updatedAt:string;
    numOfLike: number;
    numOfDislike: number;
    ownerId: string;
    minPrice: string;
    usageLimit: number;
    sellLimit: number;
    sellPrize: string;
    constructor() {
    }
}